requirements.txt
matplotlib
numpy
openai-whisper
pyaudio